# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "Health_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The API process is alive.\n\nThe Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Check whether the API process is alive",
        "tags": [
          "Health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "Health_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The API and PostgreSQL are ready.\n\nThe Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Check whether the API can reach PostgreSQL",
        "tags": [
          "Health"
        ]
      }
    },
    "/v1/availability/check": {
      "post": {
        "operationId": "Appointments_checkAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability snapshot and server-calculated end time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID or timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Unknown vehicle, dealership, or service type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check current resource availability without reserving it",
        "tags": [
          "appointments"
        ]
      }
    },
    "/v1/appointments": {
      "post": {
        "operationId": "Appointments_book",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A client-generated UUID scoped to this booking request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body or Idempotency-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Referenced record was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Slot conflict, unavailable resource, or reused idempotency key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 409,
                  "error": {
                    "code": "SLOT_CONFLICT",
                    "message": "The requested slot is no longer available"
                  },
                  "requestId": "11111111-1111-4111-8111-111111111111",
                  "timestamp": "2026-08-28T12:00:00.000Z",
                  "path": "/v1/appointments"
                }
              }
            }
          },
          "422": {
            "description": "Vehicle ownership mismatch or start time in the past",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Atomically allocate resources and book an appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/v1/appointments/{id}": {
      "get": {
        "operationId": "Appointments_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "appointments"
        ]
      }
    },
    "/v1/appointments/{id}/cancel": {
      "post": {
        "operationId": "Appointments_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled; repeat calls are idempotent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Appointment cannot be cancelled from its current status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "appointments"
        ]
      }
    }
  },
  "info": {
    "title": "Unified Service Scheduler API",
    "description": "Scenario A REST API for checking vehicle-service availability and atomically booking appointments.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Availability",
      "description": "Check technician and service-bay availability"
    },
    {
      "name": "Appointments",
      "description": "Book, retrieve, and cancel appointments"
    },
    {
      "name": "Health",
      "description": "Liveness and database readiness probes"
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "CheckAvailabilityDto": {
        "type": "object",
        "properties": {
          "dealershipId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "desiredStartAt": {
            "type": "string",
            "example": "2026-09-01T09:00:00+07:00"
          }
        },
        "required": [
          "dealershipId",
          "vehicleId",
          "serviceTypeId",
          "desiredStartAt"
        ]
      },
      "AvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean",
            "example": true
          },
          "startsAt": {
            "type": "string",
            "example": "2030-06-03T08:00:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2030-06-03T09:00:00.000Z"
          },
          "durationMinutes": {
            "type": "number",
            "example": 60
          },
          "reason": {
            "type": "string",
            "enum": [
              "OUTSIDE_BUSINESS_HOURS",
              "NO_QUALIFIED_TECHNICIAN",
              "NO_COMPATIBLE_SERVICE_BAY"
            ],
            "example": "NO_QUALIFIED_TECHNICIAN"
          }
        },
        "required": [
          "available",
          "startsAt",
          "endsAt",
          "durationMinutes"
        ]
      },
      "ErrorDetailsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SLOT_CONFLICT"
          },
          "message": {
            "type": "string",
            "example": "The requested slot is no longer available"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailsDto"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "traceId": {
            "type": "string",
            "example": "4bf92f3577b34da6a3ce929d0e0e4736"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/v1/appointments"
          }
        },
        "required": [
          "statusCode",
          "error",
          "requestId",
          "timestamp",
          "path"
        ]
      },
      "BookAppointmentDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "dealershipId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "desiredStartAt": {
            "type": "string",
            "example": "2026-09-01T09:00:00+07:00"
          }
        },
        "required": [
          "customerId",
          "vehicleId",
          "dealershipId",
          "serviceTypeId",
          "desiredStartAt"
        ]
      },
      "AppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ],
            "example": "CONFIRMED"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "dealershipId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "technicianId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceBayId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "example": "2030-06-03T08:00:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2030-06-03T09:00:00.000Z"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-28T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-08-28T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "status",
          "customerId",
          "vehicleId",
          "dealershipId",
          "serviceTypeId",
          "technicianId",
          "serviceBayId",
          "startsAt",
          "endsAt",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}